fix(rest): fix restapi plugins/sources/prebuild and plugins/sources/{name} conflict issue
